Validate wrapper CTE plans

Copilot-Session: 591f0a76-293b-4d5f-a9d8-daa91abaa286
